DNS list and create
I take the chance to restructure the project a bit. All the manual
handling of HTTP with conduit and friends is gone in favour of servant.
Library part was moved under src.

We could definitely do better in Myracloud and remove some repetition
but I want to implement few more calls first. The help text is a bit
fiddly right now and we need to set up testing.